pyproject.toml
[project]
name = "blur-script"
version = "0.1.0"
description = "壁紙用"
readme = "README.md"
authors = [
    { name = "darabsp", email = "51314762+darabsp@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
]

[project.scripts]
blur-script = "blur_script:main"

[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"
